+++ CHANGELOG.md
@@ -26,6 +26,10 @@ and this project adheres to [Semantic Versioning][].
 - The Gaussian process prior can now also be used for weights.
 - The spike and slab prior now has option to make the background distribution a Gaussian.
 - Training with sparse inputs and minibatching is about 1.5 times faster.
+- The `pl.top_weights` function now correctly sorts the features in each plot when plotting multiple factors. This is possible
+  thanks to new functionality in plotnine 0.16.
+- The `pl.weights` function now uses the entire available space for its plots, even if plotting multiple views with different
+  numbers of features. This is possible thanks to new functionality in plotnine 0.16.
 
 ### Removed
 - The MOFA compatibility mode for saving a trained model.